Художник/requirements.txt
torch
diffusers
transformers
accelerate
Pillow
gradio